moving proxy server from extress.js to lib/proxy_server.js
